Selecting a folder before downloading

+++ Fonts Downloader/Form1.cs
@@ -77,29 +77,36 @@ namespace Fonts_Downloader
         }
 
         private void CopyFont_Click(object sender, EventArgs e)
-        {          
-            //webView21.CoreWebView2.NavigateToString(reader.ReadToEnd());
-            string[] FontFileStyles = { "Thin", "ExtraLight", "Light", "Regular", "Medium", "SemiBold", "Bold", "ExtraBold", "Black" };
-            var FontWeight = new List<string>();
-            var css = new CSS();
-            var File = new FontFiles();
-            if (FontBox1.SelectedItem != null)
+        {
+            if (!string.IsNullOrEmpty(FolderName))
             {
-                string FontName = SelectedFont.Text;
-                Directory.CreateDirectory($"{FolderName}\\{FontName.Replace(" ", "")}");
-                DirectoryInfo di = new DirectoryInfo($"{FolderName}\\{FontName.Replace(" ", "")}");
-                foreach (FileInfo file in di.EnumerateFiles())
+                string[] FontFileStyles = { "Thin", "ExtraLight", "Light", "Regular", "Medium", "SemiBold", "Bold", "ExtraBold", "Black" };
+                var FontWeight = new List<string>();
+                var css = new CSS();
+                var File = new FontFiles();
+                if (FontBox1.SelectedItem != null)
                 {
-                    if (file.Name.Contains(".ttf"))
+                    string FontName = SelectedFont.Text;
+                    Directory.CreateDirectory($"{FolderName}\\{FontName.Replace(" ", "")}");
+                    DirectoryInfo di = new DirectoryInfo($"{FolderName}\\{FontName.Replace(" ", "")}");
+                    foreach (FileInfo file in di.EnumerateFiles())
                     {
-                        file.Delete();
+                        if (file.Name.Contains(".ttf"))
+                        {
+                            file.Delete();
+                        }
                     }
+                    css.CreateCSS(SizeAndStyle, SubSets, FolderName, FontName, FontFileStyle);
+                    FontWeight = css.FontWeight;
+                    File.FileLinks(SizeAndStyle, FontStyle, FontWeight, FontFileStyle, FontFileStyles, SelectedFont, FolderName, FontName, Res);
+
                 }
-                css.CreateCSS(SizeAndStyle, SubSets, FolderName, FontName, FontFileStyle);
-                FontWeight = css.FontWeight;
-                File.FileLinks(SizeAndStyle, FontStyle, FontWeight, FontFileStyle, FontFileStyles, SelectedFont, FolderName, FontName, Res);
-               
             }
+            else
+            {
+                MessageBox.Show("Please select a folder ");
+            }
+
         }
         private void webView21_CoreWebView2InitializationCompleted(object sender, Microsoft.Web.WebView2.Core.CoreWebView2InitializationCompletedEventArgs e)
         {
